# Some additional optimizations to ComponentDirectiveVisitor