--- a/src/Razor/Microsoft.AspNetCore.Razor.Language/src/RazorEngineFeatureBase.cs
+++ b/src/Razor/Microsoft.AspNetCore.Razor.Language/src/RazorEngineFeatureBase.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
@@ -32,7 +32,7 @@ namespace Microsoft.AspNetCore.Razor.Language
                 throw new InvalidOperationException(Resources.FormatFeatureMustBeInitialized(nameof(Engine)));
             }
 
-            var feature = Engine.Features.OfType<TFeature>().FirstOrDefault();
+            var feature = Engine.GetFeature<TFeature>();
             ThrowForMissingFeatureDependency<TFeature>(feature);
 
             return feature;