* Cache frequently queried metadata on TagHelperDescriptors. Determining if a tag helper is in scope
  happens very frequently and some of it's partial results can be cached on the TagHelperDescriptor instance.
* Some hash codes on immutable types - TagHelperDescriptor / TagHelperRuleDescriptor can be cached
* Remove some Linq

Turns out the benchmarks weren't running on the right inputs (the file was treated as a MVC component and was missing tag helpers).
Updated it and here's a comparison of before / after with these changes.

|                             Method |        Mean |     Error |    StdDev |     Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------- |------------:|----------:|----------:|---------:|-------:|------:|------:|----------:|
| 'TagHelper Design Time Processing' | 2,578.76 us | 46.492 us | 43.488 us |    387.8 | 3.9063 |     - |     - | 752,148 B |
| 'TagHelper Design Time Processing' |    88.77 us |  0.708 us |  0.627 us | 11,265.2 |      - |     - |     - |     576 B |

|                             Method |        Mean |     Error |    StdDev |     Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------- |------------:|----------:|----------:|---------:|-------:|------:|------:|----------:|
| 'TagHelper Design Time Processing' | 2,330.11 us | 35.816 us | 33.503 us |    429.2 | 3.9063 |     - |     - | 736,220 B |
| 'TagHelper Design Time Processing' |    64.80 us |  0.634 us |  0.593 us | 15,431.6 |      - |     - |     - |     576 B |